backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai>=1.0.0
httpx==0.26.0
PyJWT==2.8.0
cryptography==42.0.2
python-multipart==0.0.9
aiosqlite==0.19.0

# RAG / Vector Search
faiss-cpu==1.7.4
numpy>=1.24.0
